+++ packages/nuxt-localization/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/awes-io/client/compare/@awes-io/nuxt-localization@0.2.2...@awes-io/nuxt-localization@0.2.3) (2020-02-17)
+
+
+### Bug Fixes
+
+* **nuxt-localization:** show error when unauthenticated on update ([bc3e6b8](https://github.com/awes-io/client/commit/bc3e6b8679aef49d586ebe80eb6581c8fd1965a0))
+
+
+
+
+
+## [0.2.2](https://github.com/awes-io/client/compare/@awes-io/nuxt-localization@0.2.1...@awes-io/nuxt-localization@0.2.2) (2020-02-14)
+
+**Note:** Version bump only for package @awes-io/nuxt-localization
+
+
+
+
+
+## [0.2.1](https://github.com/awes-io/client/compare/@awes-io/nuxt-localization@0.2.0...@awes-io/nuxt-localization@0.2.1) (2020-02-13)
+
+**Note:** Version bump only for package @awes-io/nuxt-localization
+
+
+
+
+
 # 0.2.0 (2020-02-13)